New feature: insult random faggot in thread.

+++ 4chan.py
@@ -7,6 +7,7 @@ from string import ascii_lowercase, ascii_uppercase
 from time import sleep
 import SimpleHTTPServer, SocketServer, thread
 from datetime import datetime
+from json import loads
 
 with open("captchapage.html", "w") as capHTML:
 	fetch = urlopen("http://a.loveisover.me/dafocd.html")
@@ -43,6 +44,12 @@ def createPost(board, threadid, message, captcha, name='Anonymous'):
 	result = urlopen(url, data=data)
 	return result.getcode()
 
+def getRandomReplyFromThread(board, threadid):
+	url = 'http://a.4cdn.org/%s/thread/%s.json' % (board, threadid)
+	api = urlopen(url)
+	threadJSON = loads(api.read())
+	return threadJSON['posts'][randint(0, len(threadJSON['posts']))]['no']
+
 def hostCaptcha():
 	handler = CaptchaServer
 	serv = SocketServer.TCPServer(("", 8888), handler)
@@ -53,32 +60,52 @@ print '######################'
 print '#  AUTISM SIMULATOR  #'
 print '#        2015        #'
 print '######################'
+print '>OPTIONS'
+print '1 - SHITPOST THREAD'
+print '2 - INSULT RANDOM PEOPLE'
 print '\n'
-print 'Visit localhost:8888 for the captcha.'
+c = raw_input('YOUR CHOICE: ')
+while not c == '1' and not c == '2':
+	print '>>INVALID INPUT'
+	print '>OPTIONS'
+	print '1 - SHITPOST THREAD'
+	print '2 - INSULT RANDOM PEOPLE'
+	print '\n'
+	c = raw_input('YOUR CHOICE: ')
+
+print '\nVisit localhost:8888 for the captcha.'
 board = raw_input('Select board: (no / /)\n')
 thread = raw_input('Select thread to shitpost:\n')
 sleep(1)
 while True:
 	cap = raw_input('Enter a captcha: ')
-	cv = ['tips fedora ', 'kill yourself ', 'desu ', 'jew ', 'nigger-rigged ', 'cucked ', 'tfw ', 'ITT: ', 'memed ', 'fizz buzzed ']
-	co = ['fedora ', 'jew ', 'nigger ', 'cuck ', 'reddit ', 'tumblr ', 'bane ', 'ITT: ', 'faggot ','meme ', 'autist ', 'waifu ', 'semen demon ', 'fizz buzz ', 'trap ',': the thread ', 'doge ', 'pleb ', 'feel ']
-	com = ['a shit', 'fedora tipping ', 'an-heroing ', 'JUST ', 'jew ', 'nigger ', 'cuck ', 'based ', 'reddit-tier ', 'tumblr-tier ', 'baneposting ', 'literally ', 'faggot ','meme ', 'autistic ', 'fizz buzz-tier ', 'shit tier ', 'god tier ',':the thread ', 'pleb ', 'for free ','objectively ']
-	cg = ['Argentina is white', 'jet fuel can\'t melt steel beams','install gentoo', 'ayy lmao', 'it\'s happening', 'REEEEEEEEEEEEEEEEEEEEEEEEEEE', 'huehuehuehuehue', 'umad?', '''>>>/pol/''']
-	go = ''
-	out = ''
-	i=0; j=0; k=0
-	while i < randint(3,10):
-		out=choice(co)
-		while j < randint(2,5):
-			out+=choice(com)
-			j+=1
-		out+=choice(cv) 
-		out+= choice(co)
-		while k < randint(0,2):
-			out+=choice(com)
-			k+=1
-		go += out + '\n'
-		i+=1
-	go += choice(cg) + '.'
-	go += '\n\n/thread'
-	print "R: " + str(createPost(board, thread, go, cap))
+	if c == '1':
+		cv = ['tips fedora ', 'kill yourself ', 'desu ', 'jew ', 'nigger-rigged ', 'cucked ', 'tfw ', 'ITT: ', 'memed ', 'fizz buzzed ']
+		co = ['fedora ', 'jew ', 'nigger ', 'cuck ', 'reddit ', 'tumblr ', 'bane ', 'ITT: ', 'faggot ','meme ', 'autist ', 'waifu ', 'semen demon ', 'fizz buzz ', 'trap ',': the thread ', 'doge ', 'pleb ', 'feel ']
+		com = ['a shit', 'fedora tipping ', 'an-heroing ', 'JUST ', 'jew ', 'nigger ', 'cuck ', 'based ', 'reddit-tier ', 'tumblr-tier ', 'baneposting ', 'literally ', 'faggot ','meme ', 'autistic ', 'fizz buzz-tier ', 'shit tier ', 'god tier ',':the thread ', 'pleb ', 'for free ','objectively ']
+		cg = ['Argentina is white', 'jet fuel can\'t melt steel beams','install gentoo', 'ayy lmao', 'it\'s happening', 'REEEEEEEEEEEEEEEEEEEEEEEEEEE', 'huehuehuehuehue', 'umad?', '''>>>/pol/''']
+		go = ''
+		out = ''
+		i=0; j=0; k=0
+		while i < randint(3,10):
+			out=choice(co)
+			while j < randint(2,5):
+				out+=choice(com)
+				j+=1
+			out+=choice(cv) 
+			out+= choice(co)
+			while k < randint(0,2):
+				out+=choice(com)
+				k+=1
+			go += out + '\n'
+			i+=1
+		go += choice(cg) + '.'
+		go += '\n\n/thread'
+		print "R: " + str(createPost(board, thread, go, cap))
+	elif c == '2':
+		i = ['fucking faggot', 'retard', 'autist', 'cuck', 'jew', 'nigger', 'redditard', 'fedora',
+		'idiot', 'shitskin']
+		t = ['You goddamn %s', 'Kill yourself %s', 'OP is a %s', '%s detected', 'back to >>>/pol/ you %s']
+		reply = str(getRandomReplyFromThread(board, thread))
+		p = '>>' + reply + '\n' + choice(t) % choice(i)
+		print "R: " + str(createPost(board, thread, p, cap))
